+++ tests/bioetl/application/pipelines/test_pipeline_base.py
@@ -11,7 +11,7 @@ import pandas as pd
 import pytest
 
 from bioetl.application.pipelines.base import PipelineBase
-from bioetl.application.pipelines.contracts import ExtractorABC, LoaderABC
+from bioetl.domain.pipelines.contracts import ExtractorABC, LoaderABC
 from bioetl.application.pipelines.hooks_impl import (
     ContinueOnErrorPolicyImpl,
     FailFastErrorPolicyImpl,
